requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-nmap==0.7.1
paramiko==2.8.1
pymetasploit3==1.0.3
requests==2.26.0
python-dateutil==2.8.2 